+++ R/auc.R
@@ -348,6 +348,7 @@ pk.calc.aumc.all <- function(conc, time, ..., options=list()) {
 add.interval.col("aucinf.obs",
                  FUN="pk.calc.auc.inf.obs",
                  values=c(FALSE, TRUE),
+                 unit_type="auc",
                  desc="The area under the concentration time curve from the beginning of the interval to infinity with extrapolation to infinity from the observed Clast",
                  depends=c("lambda.z", "clast.obs"))
 PKNCA.set.summary(
@@ -360,6 +361,7 @@ PKNCA.set.summary(
 add.interval.col("aucinf.pred",
                  FUN="pk.calc.auc.inf.pred",
                  values=c(FALSE, TRUE),
+                 unit_type="auc",
                  desc="The area under the concentration time curve from the beginning of the interval to infinity with extrapolation to infinity from the predicted Clast",
                  depends=c("lambda.z", "clast.pred"))
 PKNCA.set.summary(
@@ -372,6 +374,7 @@ PKNCA.set.summary(
 add.interval.col("auclast",
                  FUN="pk.calc.auc.last",
                  values=c(FALSE, TRUE),
+                 unit_type="auc",
                  desc="The area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification")
 PKNCA.set.summary(
   name="auclast",
@@ -383,6 +386,7 @@ PKNCA.set.summary(
 add.interval.col("aucall",
                  FUN="pk.calc.auc.all",
                  values=c(FALSE, TRUE),
+                 unit_type="auc",
                  desc="The area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification plus the triangle from that last concentration to 0 at the first concentration below the limit of quantification",)
 PKNCA.set.summary(
   name="aucall",
@@ -394,6 +398,7 @@ PKNCA.set.summary(
 add.interval.col("aumcinf.obs",
                  FUN="pk.calc.aumc.inf.obs",
                  values=c(FALSE, TRUE),
+                 unit_type="aumc",
                  desc="The area under the concentration time moment curve from the beginning of the interval to infinity with extrapolation to infinity from the observed Clast",
                  depends=c("lambda.z", "clast.obs"))
 PKNCA.set.summary(
@@ -406,6 +411,7 @@ PKNCA.set.summary(
 add.interval.col("aumcinf.pred",
                  FUN="pk.calc.aumc.inf.pred",
                  values=c(FALSE, TRUE),
+                 unit_type="aumc",
                  desc="The area under the concentration time moment curve from the beginning of the interval to infinity with extrapolation to infinity from the predicted Clast",
                  depends=c("lambda.z", "clast.pred"))
 PKNCA.set.summary(
@@ -418,6 +424,7 @@ PKNCA.set.summary(
 add.interval.col("aumclast",
                  FUN="pk.calc.aumc.last",
                  values=c(FALSE, TRUE),
+                 unit_type="aumc",
                  desc="The area under the concentration time moment curve from the beginning of the interval to the last concentration above the limit of quantification")
 PKNCA.set.summary(
   name="aumclast",
@@ -429,6 +436,7 @@ PKNCA.set.summary(
 add.interval.col("aumcall",
                  FUN="pk.calc.aumc.all",
                  values=c(FALSE, TRUE),
+                 unit_type="aumc",
                  desc="The area under the concentration time moment curve from the beginning of the interval to the last concentration above the limit of quantification plus the moment of the triangle from that last concentration to 0 at the first concentration below the limit of quantification")
 PKNCA.set.summary(
   name="aumcall",