+++ R/aucint.R
@@ -229,6 +229,7 @@ pk.calc.aucint.inf.pred <- function(conc, time, start=NULL, end=NULL, time.dose,
 add.interval.col("aucint.last",
                  FUN="pk.calc.aucint.last",
                  values=c(FALSE, TRUE),
+                 unit_type="auc",
                  desc="The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast)",
                  formalsmap=list(conc="conc.group", time="time.group", time.dose=NULL))
 PKNCA.set.summary(
@@ -241,6 +242,7 @@ PKNCA.set.summary(
 add.interval.col("aucint.last.dose",
                  FUN="pk.calc.aucint.last",
                  values=c(FALSE, TRUE),
+                 unit_type="auc",
                  desc="The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast) with dose-aware interpolation/extrapolation of concentrations",
                  formalsmap=list(conc="conc.group", time="time.group", time.dose="time.dose.group"))
 PKNCA.set.summary(
@@ -253,6 +255,7 @@ PKNCA.set.summary(
 add.interval.col("aucint.all",
                  FUN="pk.calc.aucint.all",
                  values=c(FALSE, TRUE),
+                 unit_type="auc",
                  desc="The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall)",
                  formalsmap=list(conc="conc.group", time="time.group", time.dose=NULL))
 PKNCA.set.summary(
@@ -265,6 +268,7 @@ PKNCA.set.summary(
 add.interval.col("aucint.all.dose",
                  FUN="pk.calc.aucint.all",
                  values=c(FALSE, TRUE),
+                 unit_type="auc",
                  desc="The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall) with dose-aware interpolation/extrapolation of concentrations",
                  formalsmap=list(conc="conc.group", time="time.group", time.dose="time.dose.group"))
 PKNCA.set.summary(
@@ -277,6 +281,7 @@ PKNCA.set.summary(
 add.interval.col("aucint.inf.obs",
                  FUN="pk.calc.aucint.inf.obs",
                  values=c(FALSE, TRUE),
+                 unit_type="auc",
                  desc="The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast)",
                  formalsmap=list(conc="conc.group", time="time.group", time.dose=NULL),
                  depends=c("lambda.z", "clast.obs"))
@@ -290,6 +295,7 @@ PKNCA.set.summary(
 add.interval.col("aucint.inf.obs.dose",
                  FUN="pk.calc.aucint.inf.obs",
                  values=c(FALSE, TRUE),
+                 unit_type="auc",
                  desc="The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast) with dose-aware interpolation/extrapolation of concentrations",
                  formalsmap=list(conc="conc.group", time="time.group", time.dose="time.dose.group"),
                  depends=c("lambda.z", "clast.obs"))
@@ -303,6 +309,7 @@ PKNCA.set.summary(
 add.interval.col("aucint.inf.pred",
                  FUN="pk.calc.aucint.inf.pred",
                  values=c(FALSE, TRUE),
+                 unit_type="auc",
                  desc="The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall)",
                  formalsmap=list(conc="conc.group", time="time.group", time.dose=NULL),
                  depends=c("lambda.z", "clast.pred"))
@@ -316,6 +323,7 @@ PKNCA.set.summary(
 add.interval.col("aucint.inf.pred.dose",
                  FUN="pk.calc.aucint.inf.pred",
                  values=c(FALSE, TRUE),
+                 unit_type="auc",
                  desc="The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall) with dose-aware interpolation/extrapolation of concentrations",
                  formalsmap=list(conc="conc.group", time="time.group", time.dose="time.dose.group"),
                  depends=c("lambda.z", "clast.pred"))